requirements.txt
# --- Core Data Handling ---
pandas>=1.5.0
numpy>=1.23.0
pyyaml

# --- Modeling & Inference ---
# Must match the versions used during training
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib>=1.2.0
scipy>=1.10.0

# --- API & Web Server ---
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0